tools/python/airmaps/requirements.txt
omim-data-all
omim-maps_generator
apache-airflow [postgres]==1.10.10
psycopg2-binary==2.8.4
cryptography==2.8
git+https://github.com/maksimandrianov/webdav-client-python-3.git@andrianov